sanitize proposal draft text before writing to DB

+++ src/app/proposals/draft/actions/createDraftProposal.ts
@@ -5,6 +5,7 @@ import { prismaWeb2Client } from "@/app/lib/prisma";
 import { ProposalType } from "../types";
 import { DraftProposalSchema } from "../schemas/DraftProposalSchema";
 import { ProposalDraftTransaction } from "@prisma/client";
+import { sanitizeContent } from "@/lib/sanitizationUtils";
 import {
   getStageByIndex,
   getStageIndexForTenant,
@@ -19,20 +20,22 @@ const formDataByType = (
   data: z.output<typeof DraftProposalSchema>,
   id: number
 ) => {
+  // Remove the unused sanitizedTransactions variable since we're sanitizing inline
   switch (data.type) {
     case ProposalType.BASIC:
+      if (!data.transactions) return {};
       return {
         transactions: {
           // deletes old transactions so we aren't stacking on top of old transactions
           deleteMany: {},
-          create: data.transactions.map((transaction, idx) => {
+          create: data.transactions.map((transaction: { target: string; value: string; calldata: string; signature?: string; description: string }, idx: number) => {
             const asTransaction = {
               order: idx,
               target: transaction.target as string,
               value: transaction.value,
               calldata: transaction.calldata,
               signature: transaction.signature,
-              description: transaction.description,
+              description: sanitizeContent(transaction.description),
             } as ProposalDraftTransaction;
             return asTransaction;
           }),
@@ -51,11 +54,9 @@ const formDataByType = (
         social_options: {
           // deletes all existing options so we aren't stacking on top of old options
           deleteMany: {},
-          create: data.socialProposal?.options.map((option) => {
-            return {
-              text: option.text,
-            };
-          }),
+          create: data.socialProposal?.options.map((option: { text: string }) => ({
+            text: sanitizeContent(option.text),
+          })),
         },
       };
 
@@ -76,24 +77,19 @@ const formDataByType = (
           // deletes all existing options so we aren't stacking on top of old options
           // TODO: do we need to make sure deletes cascade and remove transactions?
           deleteMany: {},
-          create: data.approvalProposal.options.map((option) => {
-            return {
-              title: option.title,
-              transactions: {
-                create: option.transactions.map((transaction, idx) => {
-                  const asTransaction = {
-                    order: idx,
-                    target: transaction.target as string,
-                    value: transaction.value,
-                    calldata: transaction.calldata,
-                    description: transaction.description,
-                    proposal: { connect: { id } },
-                  };
-                  return asTransaction;
-                }),
-              },
-            };
-          }),
+          create: data.approvalProposal.options.map((option: { title: string; transactions: { target: string; value: string; calldata: string; signature?: string; description: string }[] }) => ({
+            title: sanitizeContent(option.title),
+            transactions: {
+              create: option.transactions.map((transaction: { target: string; value: string; calldata: string; signature?: string; description: string }, idx: number) => ({
+                order: idx,
+                target: transaction.target as string,
+                value: transaction.value,
+                calldata: transaction.calldata,
+                description: sanitizeContent(transaction.description),
+                proposal: { connect: { id } },
+              })),
+            },
+          })),
         },
       };
 
@@ -127,8 +123,8 @@ export async function onSubmitAction(
 
     const baseformData = {
       stage: nextStage?.stage,
-      title: parsed.data.title,
-      abstract: parsed.data.abstract,
+      title: sanitizeContent(parsed.data.title),
+      abstract: sanitizeContent(parsed.data.abstract),
       voting_module_type: parsed.data.type,
       proposal_type: parsed.data.proposalConfigType,
       proposal_scope: parsed.data.proposal_scope,
